Extract backtest replay and strategy factory modules
Phase 2 now has concrete module ownership for the pure backtest path, recorded replay path, and runtime strategy construction. The public run_strategy facade remains unchanged, and live plus recording code intentionally stay in lib.rs for the next extraction slice.

Constraint: Keep DB feed relocation and live/recording extraction as separate follow-up commits

Constraint: Preserve existing runtime mode dispatch and strategy alias behavior

Rejected: Move live and recorder code in the same commit | larger feature-cfg surface and harder rollback

Confidence: high

Scope-risk: narrow

Directive: Do not mark Phase 2 complete until live.rs, recording.rs, and DB feed relocation are handled or explicitly deferred

Tested: cargo check -p ploy-replay

Tested: cargo check -p ploy-backtest

Tested: cargo check -p ploy-strategy-runtime --no-default-features --lib

Tested: cargo test -p ploy-strategy-runtime --lib

Tested: scripts/check_feature_matrix.sh --quick

Tested: scoped rustfmt --check and git diff --check

Not-tested: live/recording extraction, DB feed relocation, database-backed backtest execution
